% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BADM}
\alias{BADM}
\title{Biomass and soil data from FluxNet sites}
\format{
## `BADM`
A data frame with 12,300 rows and 13 columns:
\describe{
  \item{SITE_ID}{Fluxnet code for the site}
  \item{LOCATION_ELEV, LOCATION_LAT, LOCATION_LON}{site coordinates}
  \item{Date}{Measurement date}
  \item{GROUP_ID}{TODO}
  \item{VARIABLE_GROUP}{category, eg abovground biomass or soil chemistry}
  \item{VARIABLE, DATAVALUE}{key and value for each measured variable}
  \item{NA_L1CODE, NA_L1NAME, NA_L2CODE, NA_L2NAME}{
		numeric IDs and names for the Level 1 and level 2 ecoregions where
		this site is located}
}
}
\source{
Originally from Fluxnet <https://fluxnet.org/badm-data-product/>,
 but the provenence and age of this specific file is not clear.
}
\usage{
BADM
}
\description{
Contains data from 246 Fluxnet sites.
Variables include aboveground and belowground biomass in various pools,
  plus soil texture/chemistry/horizonation/C&N stocks.
}
\keyword{datasets}
