% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildJAGSdata_InventoryRings.R
\name{buildJAGSdata_InventoryRings}
\alias{buildJAGSdata_InventoryRings}
\title{Format ring & plot data for JAGA}
\usage{
buildJAGSdata_InventoryRings(combined, inc.unit.conv = 0.1)
}
\arguments{
\item{combined}{object returned from matchInventoryRings. Matrix with both increment and plot data}

\item{inc.unit.conv}{conversion factor from loaded increments to cm (of radius)}
}
\value{
list
}
\description{
builds the JAGS data object for the tree ring / inventory fusion code
also sets all the priors
}
\author{
Michael Dietze
}

