% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_and_save_raster_file.R
\name{clip_and_save_raster_file}
\alias{clip_and_save_raster_file}
\title{Clip and Save a Raster File}
\usage{
clip_and_save_raster_file(
  input_path,
  polygon,
  out_path,
  mask = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{input_path}{Character. Path to the input raster file.}

\item{polygon}{An object or file coercible to a `SpatVector` by `terra::vect()`
(e.g., an `sf` object, a `SpatVector`, or a file path to a vector dataset).
used for clipping and masking. Must have a valid CRS.}

\item{out_path}{Character. Path to save the processed raster.}

\item{mask}{Logical: Should pixels outside the polygon but inside its bounding box
be masked out (TRUE) or included (FALSE)?}

\item{overwrite}{Logical: Replace output file if it already exists?}
}
\value{
Invisibly, the clipped `SpatRaster` object. The raster is also saved to `out_path`.
}
\description{
Clips a raster to a polygon bounding box, optionally masks to polygon, and saves the 
output in the same format as the input.
}
\author{
David LeBauer
}
