% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_co2_equivalent.R
\name{co2e}
\alias{co2e}
\title{Convert SOC change and greenhouse gas fluxes to CO2 equivalents}
\usage{
co2e(delta_soc = 0, ch4 = 0, n2o = 0, gwp = "AR6")
}
\arguments{
\item{delta_soc}{Numeric. Change in soil organic carbon (mass of C.)
Positive values indicate SOC gain.}

\item{ch4}{Numeric. Methane emissions as mass of CH4.}

\item{n2o}{Numeric. Nitrous oxide emissions as mass of N2O.}

\item{gwp}{Character. IPCC report containing GWP values used: "AR4", "AR5", or "AR6".}
}
\value{
Numeric. Total CO2-equivalent emissions.
}
\description{
Converts soil organic carbon (SOC) stock change and CH4/N2O fluxes to
CO2-equivalent emissions using 100-year GWP values.
}
\details{
Equations:

  CO2e_SOC = -ΔSOC * (44 / 12)

  CO2e_i = m_i * GWP100_i

GWP values:

- Default GWP100 values are from AR6 (IPCC, 2021).
- For CARB inventories, use "AR4" for comparability with current CARB inventories (CARB, 2025).
}
\examples{
co2e(delta_soc = 1)
co2e(ch4 = 1)
co2e(n2o = 1)
# return total over all sources
co2e(delta_soc = 1, ch4 = 0.1, n2o = 0.01)

}
\references{
IPCC (2021). Climate Change 2021: The Physical Science Basis (WG1 AR6).
Cambridge University Press. https://doi.org/10.1017/9781009157896

IPCC (2006). 2006 Guidelines for National Greenhouse Gas Inventories. IGES, Japan.

IPCC (2019). 2019 Refinement to the 2006 IPCC Guidelines.
IPCC, Switzerland.

Greenhouse Gas Protocol (2024). Global Warming Potential Values.
WRI/WBCSD. https://ghgprotocol.org/global-warming-potential-values

California Air Resources Board (CARB), 2025. Greenhouse Gas Global Warming Potentials. https://ww2.arb.ca.gov/ghg-gwps
}
