% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landiq_harmonized.R
\name{.prepare_pft_mapping}
\alias{.prepare_pft_mapping}
\title{Internal helper to prepare PFT mapping data}
\usage{
.prepare_pft_mapping(pft_mapping)
}
\description{
Internal helper to prepare PFT mapping data
}
\keyword{internal}
