% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.SM_CDS.R
\name{download.SM_CDS}
\alias{download.SM_CDS}
\title{Download CDS soil moisture data for the SDA workflow.}
\usage{
download.SM_CDS(
  outfolder,
  time.points,
  overwrite = FALSE,
  auto.create.key = FALSE
)
}
\arguments{
\item{outfolder}{physical paths to where the unziped soil moisture files are downloaded.}

\item{overwrite}{flag determine if we want to overwrite existing files when downloading.}

\item{auto.create.key}{flag determine if we want to automatically create the credential file.}

\item{time_points}{A vector contains each time point within the start and end date.}
}
\value{
A vector containing file paths to the downloaded files.
}
\description{
Download CDS soil moisture data for the SDA workflow.
}
\author{
Dongchen Zhang
}
