% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMIS-ET.R
\name{download_cimis_et}
\alias{download_cimis_et}
\title{Download CIMIS ETo data}
\usage{
download_cimis_et(date, local_root_dir)
}
\arguments{
\item{date}{Date to download}

\item{local_root_dir}{Root directory for storing outputs.}
}
\value{
Path to saved TIF file (invisibly)
}
\description{
Read raw ETo.asc.gz directly from CIMIS spatial portal, add CRS, and save
locally as Cloud-optimized GeoTIFF. Outputs will be saved to
`<local_root_dir>/CIMIS-ETo-YYYY-MM-DD.tif`.
}
