% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NEON_multi_site.R
\name{Download_multiSites_NEON}
\alias{Download_multiSites_NEON}
\title{Title generalized for download NEON files of multiSites. and automatically detect the most recent available NEON data and if data is problematic, try next year.}
\usage{
Download_multiSites_NEON(settings, main_store_dir)
}
\arguments{
\item{settings}{a multisite settings object, should work with a single site setting.}

\item{main_store_dir}{the preferred path to store the downloaded NEON files. the log file and VEG_INFO should also be stored in that path}
}
\value{
VEG_INFO: a list object containing veg_info, site IDs of each site.
}
\description{
Title generalized for download NEON files of multiSites. and automatically detect the most recent available NEON data and if data is problematic, try next year.
}
\examples{
main_store_dir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/NEON42/IC/NEON_downloads"
          settings <- read.settings("/projectnb/dietzelab/dongchen/All_NEON_SDA/NEON42/pecan.xml")
          vege_info <- Download_multiSites_NEON(settings, main_store_dir)
}
