% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataONE_doi_download.R
\name{download_package_rm}
\alias{download_package_rm}
\title{download_packages}
\usage{
download_package_rm(resource_map, directory, CNode = "PROD",
  download_format = "application/bagit-097",
  overwrite_directory = TRUE)
}
\arguments{
\item{resource_map}{the resource map that corresponds to the given data package}

\item{directory}{location that download.packages places the data}

\item{CNode}{defaults to "PROD"}

\item{download_format}{typically "application/bagit-097". Other possible formats currently unknown.}

\item{overwrite_directory}{boolean that indicates whether or not the function should overwrite the directory}
}
\value{
results of download
}
\description{
Uses resource_map and dataone::getPackage to download the data into a BagItFile. Then utils::unzip unzips the data and stores in the user's directory.
}
