% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vswc.R
\name{download_vswc}
\alias{download_vswc}
\title{download_vswc}
\usage{
download_vswc(outdir, site_ID)
}
\arguments{
\item{outdir}{location where you want the soil moisture ncdf file saved}

\item{site_ID}{the NEON site ID}
}
\value{
a ncdf file with the soil moisture value for the most recent available day
}
\description{
Download Soil Moisture (VSWC)
}
\examples{
outdir <- "/projectnb/dietzelab/cfranci1/soilmoist" site_ID <- "BART"
}
\author{
Christina Francis and Alexis Helgeson
}
