% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens.veg.module.R
\name{ens_veg_module}
\alias{ens_veg_module}
\title{Sampling/ensemble module}
\usage{
ens_veg_module(getveg.id, dbparms, input_veg, outfolder, machine,
  start_date, end_date, n.ensemble, new_site, host)
}
\arguments{
\item{getveg.id}{list, input.id and dbfile.id of the IC file in intermediate pecan standard}

\item{dbparms}{list, settings$database info reqired for opening a connection to DB}

\item{input_veg}{list, this is a sublist of settings$run$inputs that has info about source, id, metadata of the requested IC file}

\item{outfolder}{path to where the processed files will be written}

\item{machine}{data frame, DB info regarding localhost machine id/hostname etc.}

\item{start_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB}

\item{end_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB}

\item{n.ensemble}{integer, ensemble member number}

\item{new_site}{data frame, id/lat/lon/name info about the site}

\item{host}{list, host info as in settings$host, host$name forced to be "localhost" upstream}
}
\description{
Sampling/ensemble module
}
\author{
Istem Fer
}
