% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_params_extraction.R
\name{estimate_dirichlet_parameters}
\alias{estimate_dirichlet_parameters}
\title{A function to estimate individual alphas for Dirichlet distribution to approximate the observed quantiles with means as known moments}
\usage{
estimate_dirichlet_parameters(means, quantiles)
}
\arguments{
\item{means}{A vector of means of sand, clay, and silt proportion for one soil layer at one site from SoilGrids data}

\item{quantiles}{A list of 5th, 50th, and 95th percentiles for sand, clay, and silt for one soil layer at one site from SoilGrids data}
}
\description{
A function to estimate individual alphas for Dirichlet distribution to approximate the observed quantiles with means as known moments
}
