% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_to_etc.R
\name{eto_to_etc}
\alias{eto_to_etc}
\title{Convert reference evapotranspiration to crop evapotranspiration}
\usage{
eto_to_etc(eto, kc)
}
\arguments{
\item{eto}{numeric vector of reference evapotranspiration}

\item{kc}{numeric vector of crop coefficients}
}
\value{
numeric vector of crop evapotranspiration
}
\description{
Multiply reference evapotranspiration (ETo) by a unitless
crop coefficient (Kc) to produce crop evapotranspiration (ETc).
Output has the same units as input ETo, with length/time scale,
e.g., mm d-1.
}
