% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.stringCode.R
\name{extract.stringCode}
\alias{extract.stringCode}
\title{extract.stringCode}
\usage{
extract.stringCode(x, extractor = from.TreeCode)
}
\arguments{
\item{x}{string to decode}

\item{extractor}{function to apply}
}
\description{
extract.stringCode
}
