% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMIS-ET.R
\name{extract_cimis_date}
\alias{extract_cimis_date}
\title{Extract CIMIS daily reference ETo values}
\usage{
extract_cimis_date(
  design_points,
  date,
  local_root_dir,
  download_missing = FALSE
)
}
\arguments{
\item{design_points}{`data.frame` of design points with columns
`location_id`, `lat`, and `lon`}

\item{date}{Date to download}

\item{local_root_dir}{Root directory for storing outputs.}

\item{download_missing}{If `TRUE` and the local COG is missing, download it.
If `FALSE` and the file is missing, throw an error.}
}
\value{
`design_points` `data.frame` with additional columns `date`, and
`etref_mm_day` (reference ET, mm/day)
}
\description{
Extract CIMIS daily reference ETo values
}
