% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_FIA.R
\name{extract_FIA}
\alias{extract_FIA}
\title{extract_FIA}
\usage{
extract_FIA(lon, lat, site_name, start_date, end_date, gridres = 0.075,
  dbparms, outfolder, overwrite = FALSE, ...)
}
\author{
Istem Fer
}
