% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openET.R
\name{extract_openet_daily}
\alias{extract_openet_daily}
\title{Extract daily ET data from OpenET}
\usage{
extract_openet_daily(design_points, start_date, end_date)
}
\arguments{
\item{design_points}{`data.frame` of design points with columns `lat` and `lon`}

\item{start_date}{Start date for data extraction}

\item{end_date}{End date for data extraction}
}
\value{
`design_points` `data.frame` with additional columns `date`, and
`et_mm_day` (ET, mm/day)
}
\description{
Note that this requires the environment variable `OPENET_API_KEY` to be set.
A convenient way to do this is via a `.Renviron`, either globally
(`~/.Renviron`) or in the current working directory (`./.Renviron`), with
contents like:
}
\details{
```
OPENET_API_KEY="abcdefg123456"
```

You can obtain an OpenET API key from the OpenET data portal.
}
