% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_SM_CDS.R
\name{extract_SM_CDS}
\alias{extract_SM_CDS}
\title{Extract CDS soil moisture data for the SDA workflow.}
\usage{
extract_SM_CDS(
  site_info,
  time.points,
  in.path,
  out.path = NULL,
  allow.download = TRUE,
  search_window = 10
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{time.points}{A vector contains each time point within the start and end date.}

\item{in.path}{physical paths to where the unziped soil moisture files are downloaded.}

\item{out.path}{Where the final CSV file will be stored.}

\item{allow.download}{Flag determine if we want to automatic download files if they are not available.}

\item{search_window}{time length (days) for locate available soil moisture values.}
}
\value{
A data frame containing soil moisture and sd for each site and each time step.
}
\description{
Extract CDS soil moisture data for the SDA workflow.
}
\author{
Dongchen Zhang
}
