% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{extract_soil_gssurgo}
\alias{extract_soil_gssurgo}
\title{Extract soil data from gssurgo}
\usage{
extract_soil_gssurgo(
  outdir,
  lat,
  lon,
  aoi = NULL,
  size = 1,
  radius = 500,
  depths = c(0.15, 0.3, 0.6)
)
}
\arguments{
\item{outdir}{Output directory for writing down the netcdf file}

\item{lat}{Latitude of center point (single numeric value)}

\item{lon}{Longitude of center point (single numeric value)}

\item{aoi}{Spatial polygon defining area of interest.
When this is specified, all three of `lat`, `lon`, and `radius` are ignored}

\item{size}{Ensemble size}

\item{radius}{distance in meters to sample around center point}

\item{depths}{depths in m of boundaries between soil layers}
}
\value{
list of paths to the generated soil netcdf files
}
\description{
Samples soil variabity within an area specified either as a radius around a
single lat/lon point OR as a spatial polygon.
}
\details{
The samples are drawn from all soil types present in the area of interest,
as determined by the gSSURGO soil map at 30 meter pixel resolution.
Each soil component is sampled proportionally to its prevalence in the AOI,
with the restriction that each soil type is sampled at least once.
Note especially that this means sites with multiple soil types will always
have multiple output files, **even when** `size = 1`.
}
\examples{
\dontrun{
   outdir  <- "~/paleon/envTest"
   lat     <- 40
   lon     <- -80
   PEcAn.data.land::extract_soil_gssurgo(outdir, lat, lon)
}
}
\author{
Hamze Dokoohaki, Akash
}
