% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_veg.R
\name{extract_veg}
\alias{extract_veg}
\title{extract_veg}
\usage{
extract_veg(
  new_site,
  start_date,
  end_date,
  source,
  gridres,
  format_name = NULL,
  machine_host,
  dbparms,
  outfolder,
  overwrite = FALSE,
  input_veg = input_veg,
  ...
)
}
\arguments{
\item{new_site}{new_site object passed from ic_process includes lat, lon, id, and name}

\item{start_date}{"YYYY-MM-DD"}

\item{end_date}{"YYYY-MM-DD"}

\item{source}{taken from input$source, passed from ic_process}

\item{gridres}{only used for source = "FIA"}

\item{format_name}{DEFAULT=NULL}

\item{machine_host}{passed from ic_process}

\item{dbparms}{taken from settings object, passed from ic_process}

\item{outfolder}{passed from ic_process, location where to store files}

\item{overwrite}{DEFAULT = FALSE}

\item{input_veg}{passed from input object in ic_process}

\item{...}{Additional parameters}
}
\value{
results object to be passed back to get.veg.module
}
\description{
Function queries a DB to extract veg info downstream
}
\author{
Istem Fer and Alexis Helgeson
}
