% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_components.R
\name{fertilizer_components}
\alias{fertilizer_components}
\title{Calculate the Nitrogen and Carbon Content of a Fertilizer Application}
\usage{
fertilizer_components(
  fertilizer_type,
  fertilizer_amount,
  fraction_organic_n = NULL,
  fraction_organic_c = NULL,
  cn_ratio = NULL
)
}
\arguments{
\item{fertilizer_type}{Character string specifying the type of fertilizer. Can be one of "urea", "ammonium_nitrate", "compost", "manure", etc., or a string in the format "NN-PP-KK" (e.g., "45-5-10").}

\item{fertilizer_amount}{Numeric value specifying the amount of fertilizer applied in kg/ha.}

\item{fraction_organic_n}{Optional numeric value specifying the fraction of nitrogen in the fertilizer. Used for organic fertilizers if not provided in the dataset.}

\item{fraction_organic_c}{Optional numeric value specifying the fraction of carbon in the fertilizer. Used for organic fertilizers if not provided in the dataset.}
}
\value{
A list containing:
  - `fertilizer_type`: The type of fertilizer used.
  - `NO3_N`: The amount of nitrate nitrogen (NO3-N) in kg/ha.
  - `NH4_N`: The amount of ammonium nitrogen (NH4-N) in kg/ha.
  - `N_org`: The amount of organic nitrogen in kg/ha.
  - `C_org`: The amount of organic carbon in kg/ha.
}
\description{
This function calculates the different forms of nitrogen (NO3-N, NH4-N, organic N) and organic carbon (C_org) in a fertilizer application.
It can determine fertilizer nitrogen and carbon content using either a lookup table based on 
the SWAT model's [`fertilizer.frt`](https://github.com/swat-model/swatplus/blob/main/data/Osu_1hru/fertilizer.frt)
file or determine the fertilizer's 
nutrient content based on NN-PP-KK format.
}
\details{
Consistent with assumptions (TODO: confirm this statement) in DayCent, DSSAT, and other models, urea is treated as NH3 because the transformation typically occurs within a day.
}
\examples{
fertilizer_components("45-00-00", 200)
fertilizer_components("compost", 1000)
fertilizer_components("manure", 1000, fraction_organic_n = 0.02, fraction_organic_c = 0.08)

}
