% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia2ED.R
\name{fia.to.psscss}
\alias{fia.to.psscss}
\title{fia.to.psscss}
\usage{
fia.to.psscss(settings, lat = as.numeric(settings$run$site$lat),
  lon = as.numeric(settings$run$site$lon),
  year = lubridate::year(settings$run$start.date), gridres = 0.075,
  min.year = year - 5, max.year = year + 5, overwrite = FALSE)
}
\arguments{
\item{create}{pss/css files based on data in the fia database}
}
\value{
nothing
}
\description{
convert x into a table
}
\author{
Mike Dietze, Rob Kooper, Ryan Kelly
}
