% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.land}
\alias{find.land}
\title{find.land}
\usage{
find.land(lat, lon, plot = FALSE)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}
}
\value{
data frame with numeric lat, lon and logical value "land"
}
\description{
find land
}
\details{
extract terrestrial lat and lon coordinates from a grid
}
\author{
David LeBauer
}

