% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataONE_doi_download.R
\name{format_identifier}
\alias{format_identifier}
\title{Functions to determine if data can be found by doi in R
Author: Liam Burke
Code draws heavily on dataoneR package for communication with the DataONE federation 
format.identifier}
\usage{
format_identifier(id)
}
\arguments{
\item{id}{the doi or other identifier linked to the package in DataONE}
}
\value{
returns the id in the proper format for querying the DataONE Federation (using solrQuery syntax)
}
\description{
Functions to determine if data can be found by doi in R
Author: Liam Burke
Code draws heavily on dataoneR package for communication with the DataONE federation 
format.identifier
}
\author{
Liam P Burke, \email{lpburke@bu.edu}
}
