% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_SOILGRID_Utilities.R
\name{generate_soilgrids_ensemble}
\alias{generate_soilgrids_ensemble}
\title{Generate ensemble members for a site}
\usage{
generate_soilgrids_ensemble(
  processed_data,
  site_id,
  lat,
  lon,
  ensemble_size,
  verbose = FALSE
)
}
\arguments{
\item{processed_data}{Processed soil carbon data}

\item{site_id}{Site ID}

\item{lat}{Site latitude}

\item{lon}{Site longitude}

\item{ensemble_size}{Number of ensemble members to create}

\item{verbose}{Print detailed progress information to the terminal? TRUE/FALSE}
}
\value{
List of ensemble data for the site
}
\description{
Generate ensemble members for a site
}
