\name{get.attributes}
\alias{get.attributes}
\title{Retrieve attribute information from a vector or raster layer}
\usage{
  get.attributes(file, coords)
}
\arguments{
  \item{file}{vector or raster layer}

  \item{coords}{dataframe containing coordinates of a
  region}
}
\description{
  Function to extract attribute information from vector or
  raster data layer.
}
\author{
  Shawn P. Serbin
}

