% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.soil.R
\name{get.soil}
\alias{get.soil}
\title{get.soil}
\usage{
get.soil(lat, lon, soil.nc = soil.nc)
}
\arguments{
\item{soil.nc}{netCDFe file with soil data}

\item{latitude}{}

\item{longitude}{}
}
\value{
usda soil class
}
\description{
Get Soil
}
\author{
David LeBauer
}

