% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.functions.R
\name{get_attributes}
\alias{get_attributes}
\alias{get_attributes,}
\alias{get.attributes}
\title{Retrieve Attribute Information from Vector or Raster Layers}
\usage{
get_attributes(file, coords)
}
\arguments{
\item{file}{Character. Path to the vector (e.g., Shapefile, GeoPackage) or raster (e.g., GeoTIFF) file.}

\item{coords}{Numeric vector. A vector containing `xmin`, `ymin`, `xmax`, `ymax` defining the bounding box for subsetting.}
}
\value{
- For vector files: An `sf` object containing the subset of features within the bounding box.
- For raster files: A `SpatRaster` object cropped to the bounding box.
}
\description{
This function extracts attribute information from a vector or raster file within a specified bounding box defined by coordinates.
}
\examples{
\dontrun{
# Example with a vector file (Shapefile)
vector_file <- "path/to/vector.shp"
bbox_coords <- c(-95, 42, -84, 47) # xmin, ymin, xmax, ymax
subset_vector <- get_attributes(file = vector_file, coords = bbox_coords)
print(subset_vector)

# Example with a raster file (GeoTIFF)
raster_file <- "path/to/raster.tif"
subset_raster <- get_attributes(file = raster_file, coords = bbox_coords)
print(subset_raster)
}
}
\author{
Shawn P. Serbin

David LeBauer
}
