% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_parquet_to_json.R
\name{get_event_ensemble_ids}
\alias{get_event_ensemble_ids}
\title{Efficiently extract unique event IDs from parquet files}
\usage{
get_event_ensemble_ids(event_paths, parquet_dir = NULL)
}
\arguments{
\item{event_paths}{(character) Named character vector of paths to PEcAn
standard event parquet files. Names correspond to event types; e.g.,
`c(leafon = "/path/to/leaf_on.parquet")`. If not names are provided, they
are deduced from the file name (stripping the `.parquet` extension). Folders
containing parquet files are also supported.}

\item{parquet_dir}{Alternative way to pass `event_paths` by just specifying
a directory of PEcAn standard event files. Each file or folder must be named
`<event_type>.parquet` (e.g., `irrigation.parquet/`, `leafon.parquet`).}
}
\value{
Named list of unique ensemble IDs for each event type. Names
correspond to event types.
}
\description{
Efficiently extract unique event IDs from parquet files
}
