% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-landiq.R
\name{get_landiq}
\alias{get_landiq}
\title{Get LandIQ parcels and crop data}
\usage{
get_landiq(design_points, parcels_file, crops_file)
}
\arguments{
\item{design_points}{`data.frame` of coordinates to extract. Must contain
columns `id`, `lat`, and `lon`.}

\item{parcels_file}{Path to harmonized LandIQ parcels (GPKG) file}

\item{crops_file}{Path to LandIQ crops parquet file}
}
\value{
`design_points` `data.frame` with LandIQ parcel IDs, year, season, CLASS, and SUBCLASS
}
\description{
Get LandIQ parcels and crop data
}
