% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-landiq-parcels.R
\name{get_landiq_parcel_ids}
\alias{get_landiq_parcel_ids}
\title{Get Parcel IDs from LandIQ}
\usage{
get_landiq_parcel_ids(design_points, parcels_file)
}
\arguments{
\item{design_points}{`data.frame` of coordinates to extract. Must contain
columns `id`, `lat`, and `lon`.}

\item{parcels_file}{Path to harmonized LandIQ parcels (GPKG) file}
}
\value{
`design_points` `data.frame` with harmonized LandIQ parcel_IDs
}
\description{
Get Parcel IDs from LandIQ
}
