% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataONE_doi_download.R
\name{get_resource_map}
\alias{get_resource_map}
\title{get.resource.map}
\usage{
get_resource_map(id, CNode = "PROD")
}
\arguments{
\item{id}{the doi or other identifier linked to the package in DataONE}

\item{CNode}{default is "PROD"}
}
\value{
return the resource_map or a message indicating that there is no corresponding resource_map for the given id
}
\description{
get.resource.map
}
