% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.veg.module.R
\name{get_veg_module}
\alias{get_veg_module}
\title{Load/extract + match species module}
\usage{
get_veg_module(input_veg, outfolder, start_date, end_date, dbparms,
  new_site, host, machine_host, overwrite)
}
\arguments{
\item{input_veg}{list, this is a sublist of settings$run$inputs that has info about source, id, metadata of the requested IC file}

\item{outfolder}{path to where the processed files will be written}

\item{start_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB}

\item{end_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB}

\item{dbparms}{list, settings$database info reqired for opening a connection to DB}

\item{new_site}{data frame, id/lat/lon/name info about the site}

\item{host}{list, host info as in settings$host, host$name forced to be "localhost" upstream}

\item{machine_host}{local machine hostname, e.g. "pecan2.bu.edu"}

\item{overwrite}{logical flag for convert.input}
}
\description{
Load/extract + match species module
}
\author{
Istem Fer
}
