% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{gSSURGO.Query}
\alias{gSSURGO.Query}
\title{Extract soil data from gSSURGO database}
\usage{
gSSURGO.Query(mukeys)
}
\arguments{
\item{mukeys}{}

}
\description{
Extract soil data from gSSURGO database
}
\examples{

}
