% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{IC_Maker}
\alias{IC_Maker}
\title{IC_Maker}
\usage{
IC_Maker(settings, ens.n = 5, outdir)
}
\arguments{
\item{settings}{PEcAn xml settings}

\item{ens.n}{number of ensemble memebers needs to be genrated}

\item{outdir}{output directory}
}
\value{
xml setting
}
\description{
This function generates a series of ensemble members for initial condition using the BADM database.
}
\examples{
\dontrun{
settings <- PEcAn.settings::read.settings("pecan.SDA.10sites.xml")

 suppressMessages({
  setting.with.ic <- PEcAn.settings::papply(settings,
                          IC_Maker,
                          ens.n=10,
                          outdir="/fs/data3/hamzed/MultiSite_Project/IC/ICFiles")
})
} 
}
