% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_status_check.R
\name{ic_status_check}
\alias{ic_status_check}
\title{ic_status_check}
\usage{
ic_status_check(getveg.id = NULL, putveg.id = NULL, input, site_id, model,
  start_date, con)
}
\description{
Function to check whether there are processed ic files already
}
\author{
Istem Fer
}
