% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataONE_doi_download.R
\name{id.resolveable}
\alias{id.resolveable}
\title{id.resolveable}
\usage{
id.resolveable(id, return_result = TRUE, CNode = "PROD")
}
\arguments{
\item{id}{the doi or other identifier linked to the package in DataONE}

\item{return_result}{boolean that returns or suppresses result of query. defaults to TRUE.}

\item{CNode}{CNode="PROD"}
}
\value{
returns message indicating wether or not the id resolves to data in the DataONE federation
}
\description{
id.resolveable
}
