% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryGrowthFusionDiagnostics.R
\name{InventoryGrowthFusionDiagnostics}
\alias{InventoryGrowthFusionDiagnostics}
\title{InventoryGrowthFusionDiagnostics}
\usage{
InventoryGrowthFusionDiagnostics(jags.out, combined)
}
\arguments{
\item{jags.out}{output mcmc.list from InventoryGrowthFusion}

\item{combined}{data output from matchInventoryRings}
}
\author{
Michael Dietze
}

