% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iscn_soc}
\alias{iscn_soc}
\title{Soil organic carbon (SOC) density based on ecoregion level 2 code from the ISCN database.}
\format{

}
\source{
https://iscn.fluxdata.org/wp-content/uploads/sites/23/2019/05/ISCN_ALL_DATA_DATASET_1-1.xlsx
}
\usage{
iscn_soc
}
\description{
Contains 200 ensemble SOC data from 43 level 2 ecoregions across North America.
Variable include SOC densities in g/cm2.
}
\keyword{datasets}
