% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landiq_crop_mapping_codes}
\alias{landiq_crop_mapping_codes}
\title{LandIQ crop mapping codes}
\format{
## `landiq_crop_mapping_codes`
A data frame with 203 rows and 4 columns:
\describe{
  \item{CLASS}{LandIQ class code.}
  \item{class_name}{LandIQ class name.}
  \item{SUBCLASS}{LandIQ subclass code.}
  \item{subclass_name}{LandIQ subclass name.}
}
}
\source{
California Department of Water Resources. (2023). Statewide Crop Mapping—California
Natural Resources Agency Open Data. Metadata retrieved from https://data.cnra.ca.gov/dataset/statewide-crop-mapping and manually extracted into `data-raw/landiq_crop_mapping_codes.tsv`.
}
\usage{
landiq_crop_mapping_codes
}
\description{
LandIQ land-use class and subclass labels used for crop mapping.
}
\keyword{datasets}
