% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_veg.R
\name{load_veg}
\alias{load_veg}
\title{load_veg}
\usage{
load_veg(
  new_site,
  start_date,
  end_date,
  source_id,
  source,
  icmeta = NULL,
  format_name = NULL,
  machine_host,
  dbparms,
  outfolder,
  overwrite = FALSE,
  ...
)
}
\description{
uses `PEcAn.benchmark::load_data()` to get veg data
}
\author{
Istem Fer
}
