% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_veg_manual.R
\name{load_veg_manual}
\alias{load_veg_manual}
\title{load_veg_manual}
\usage{
load_veg_manual(
  new_site,
  start_date,
  end_date,
  source_id,
  source,
  icmeta = NULL,
  format_name = NULL,
  machine_host,
  dbparms,
  outfolder,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{new_site}{data frame, id/lat/lon/name info about the site}

\item{start_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB}

\item{end_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB}

\item{source_id}{taken from input_veg. Numeric ID of the input in the PEcAn database}

\item{source}{taken from input_veg. The input data type. This tag name needs to match the names in the corresponding conversion functions. If you are using PEcAn’s automatic input processing, this is the only field you need to set. However, this field is ignored if id and/or path are provided.}

\item{icmeta}{default NULL. Info taken from settings$run$inputs$css$metadata$area.}

\item{format_name}{default NULL. Info taken from format$vars$bety_name}

\item{machine_host}{local machine hostname, e.g. "pecan2.bu.edu"}

\item{dbparms}{list, settings$database info reqired for opening a connection to DB}

\item{outfolder}{path to where the processed files will be written}

\item{overwrite}{Default is FALSE. logical flag for convert.input}

\item{...}{Additional parameters}
}
\description{
Function uses load_data{benchmark} to get veg data
}
\author{
Istem Fer
}
