% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInventoryRings.R
\name{matchInventoryRings}
\alias{matchInventoryRings}
\title{matchInventoryRings}
\usage{
matchInventoryRings(
  trees,
  rings,
  extractor = "TreeCode",
  nyears = 30,
  coredOnly = TRUE
)
}
\arguments{
\item{trees, rings}{codes from which to extract IDs}

\item{extractor}{function to call, specified without its initial `to.`
e.g. "TreeCode" calls `to.TreeCode`}

\item{nyears}{number of years to extract}

\item{coredOnly}{logical: Only include trees with data from 2000?}
}
\description{
matchInventoryRings
}
