% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mslsp_to_canopycover.R
\name{mslsp_to_canopycover}
\alias{mslsp_to_canopycover}
\title{Convert MSLSP phenology data to tidy canopy cover}
\usage{
mslsp_to_canopycover(mslsp_path, parcel_ids = NULL, years = NULL)
}
\arguments{
\item{mslsp_path}{Path to directory containing MSLSP outputs (in parquet format)}

\item{parcel_ids}{Vector of parcel IDs for filtering. If `NULL`, use all parcels.}

\item{years}{Vector of years for filtering. If `NULL`, use all years.}
}
\value{
`data.frame` of `parcel_id`, `year`, `season`, `date`, and
`canopy_cover` (fraction). `date` is a sequence from the MSLSP greenness
onset (growing season start) to the greenness minimum (growing season end).
`canopy_cover` is the fractional canpoy cover (0 to 1), suitable for ingest
into [eto_to_etc_bism()] in "canopy cover mode".
}
\description{
Convert MSLSP phenology data to tidy canopy cover
}
