% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{netcdf.writer.BADAM}
\alias{netcdf.writer.BADAM}
\title{netcdf.writer.BADAM}
\usage{
netcdf.writer.BADAM(lat, long, siteid, outdir)
}
\arguments{
\item{lat}{numeric latitude}

\item{long}{numeric longitude}

\item{siteid}{site id as a string}

\item{outdir}{outputdir which you want to store the IC netcdf file}
}
\value{
a string of the address where the IC file is stored
}
\description{
netcdf.writer.BADAM
}
