% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids_soc_extraction.R
\name{ocs_extract}
\alias{ocs_extract}
\title{ocs_extract}
\usage{
ocs_extract(site_info, outdir = NULL, verbose = TRUE, ...)
}
\arguments{
\item{site_info}{A dataframe of site info containing the BETYdb site ID, 
site name, latitude, and longitude, e.g. 
(site_id, site_name, lat, lon)}

\item{outdir}{Optional. Provide the results as an Rdata file 
(soilgrids_soc_data.RData)}

\item{verbose}{Provide progress feedback to the terminal? TRUE/FALSE}
}
\description{
ocs_extract function
A function to extract total soil organic carbon for a single or group of 
lat/long locationsbased on user-defined site location from SoilGrids250m 
version 2.0 : https://soilgrids.org
}
\author{
Qianyu Li, Shawn P. Serbin
}
