% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_utils.R
\name{om2soc}
\alias{om2soc}
\title{Convert organic matter to soil organic carbon}
\usage{
om2soc(om_percent)
}
\arguments{
\item{om_percent}{organic matter percentage (0-100)}
}
\value{
soil organic carbon percentage (0-100)
}
\description{
Converts organic matter content to soil organic carbon using the Van Bemmelen factor (1.724).
}
\author{
Akash
}
