% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.MatrixNames.R
\name{parse.MatrixNames}
\alias{parse.MatrixNames}
\title{parse.MatrixNames}
\usage{
parse.MatrixNames(w, pre = "x", numeric = FALSE)
}
\arguments{
\item{w}{mcmc object containing matrix outputs}

\item{pre}{prefix (variable name) for the matrix variable to be extracted}

\item{numeric}{boolean, whether to coerce class to numeric}
}
\value{
matrix
}
\description{
parse.MatrixNames
}
\author{
Michael Dietze
}

