% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_roots.R
\name{partition_roots}
\alias{partition_roots}
\title{partition_roots}
\usage{
partition_roots(roots, rtsize)
}
\arguments{
\item{roots}{vector of root carbon values in kg C m-2}

\item{rtsize}{vector of lower bounds of root size class thresholds in m, length greater than one and equal to roots. Must contain threshold within .0005 m of .002 m}
}
\value{
list containing summed fine root and coarse root carbon (2 values)
}
\description{
Given a vector of root size thresholds (lower bound of each) and a vector of corresponding root carbon values, partition_roots checks if the input can be partitioned along the .002 m threshold between fine and coarse roots and returns a list containing the summed values for fine and coarse. If there are fewer than two thresholds or none within .0005 m of .002 m, returns NULL. Meant to be used in conjunction with standard variable root_carbon_content with rtsize dimension, extracted from netcdf.
}
\author{
Anne Thomas
}
