% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_ic_list2netcdf.R
\name{pool_ic_list2netcdf}
\alias{pool_ic_list2netcdf}
\title{pool_ic_list2netcdf}
\usage{
pool_ic_list2netcdf(input, outdir, siteid, ens = NA)
}
\arguments{
\item{input}{list with two elements: list of netcdf dimensions (dims, with named values) and list of variables (vals, with named values)}

\item{outdir}{directory to write netcdf file}

\item{siteid}{site id}

\item{ens}{Default is NA. Ensemble members.}
}
\description{
Converts input list containing standard dimensions and variables (named values) for initial conditions to a netcdf file, input to pool-based models.
}
\author{
Anne Thomas
}
