% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_ic_netcdf2list.R
\name{pool_ic_netcdf2list}
\alias{pool_ic_netcdf2list}
\title{pool_ic_netcdf2list}
\usage{
pool_ic_netcdf2list(nc.path)
}
\arguments{
\item{nc.path}{path to netcdf file containing standard dimensions and variables}
}
\value{
list with two elements: list of netcdf dimensions (dims, with named values) and list of variables (vals, with named values)
}
\description{
Converts netcdf containing standard dimensions and variables for pool-based initial conditions, created by pool_ic_list2netcdf, back into list format
}
\author{
Anne Thomas
}
