% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_SOILGRID_Utilities.R
\name{preprocess_soilgrids_data}
\alias{preprocess_soilgrids_data}
\title{Preprocess SoilGrids data for ensemble generation}
\usage{
preprocess_soilgrids_data(soil_data, verbose = FALSE)
}
\arguments{
\item{soil_data}{Dataframe with SoilGrids soil carbon data}

\item{verbose}{Logical, print detailed progress information}
}
\value{
List containing processed data and CV distributions for both depths
}
\description{
Preprocess SoilGrids data for ensemble generation
}
