% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.veg.module.R
\name{put_veg_module}
\alias{put_veg_module}
\title{Match species to PFTs + veg2model module}
\usage{
put_veg_module(getveg.id, dbparms, input_veg, pfts, outfolder, n.ensemble,
  dir, machine, model, start_date, end_date, new_site, host, overwrite)
}
\arguments{
\item{getveg.id}{list, input.id and dbfile.id of the IC file in intermediate pecan standard}

\item{dbparms}{list, settings$database info reqired for opening a connection to DB}

\item{input_veg}{list, this is a sublist of settings$run$inputs that has info about source, id, metadata of the requested IC file}

\item{pfts}{list, same as settings$pfts}

\item{outfolder}{path to where the processed files will be written}

\item{n.ensemble}{integer, ensemble member number}

\item{dir}{dir path to dbfiles on local machine}

\item{machine}{data frame, DB info regarding localhost machine id/hostname etc.}

\item{model}{model name, e.g. "ED2"}

\item{start_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB}

\item{end_date}{date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB}

\item{new_site}{data frame, id/lat/lon/name info about the site}

\item{host}{list, host info as in settings$host, host$name forced to be "localhost" upstream}

\item{overwrite}{logical flag for convert.input}
}
\description{
Match species to PFTs + veg2model module
}
\author{
Istem Fer
}
