% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPNET_flux_gapfill.R
\name{read_pred_data}
\alias{read_pred_data}
\title{read_pred_data}
\usage{
read_pred_data(pred.var.dir, site_ids)
}
\arguments{
\item{pred.var.dir}{character: path to the predictor/flux data file.}

\item{site_ids}{character: vector of site IDs to retain.}
}
\value{
data.frame containing filtered predictor/flux data.
}
\description{
This function reads the predictor/flux dataset and keeps only rows
corresponding to the selected site IDs.
}
\author{
Yang Gu
}
