% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPNET_flux_gapfill.R
\name{read_site_ids}
\alias{read_site_ids}
\title{read_site_ids}
\usage{
read_site_ids(site.dir, n_sites = 40)
}
\arguments{
\item{site.dir}{character: path to the site information table.}

\item{n_sites}{numeric: number of sites to keep from the site table,
the total number of sites is 241.}
}
\value{
character vector of selected site IDs. Example: 'Site ID' CA-ARB
}
\description{
This function reads the site information table and returns the selected
site IDs for downstream modeling.
}
\author{
Yang Gu
}
