% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_utils.R
\name{remove_dead_trees}
\alias{remove_dead_trees}
\title{remove_dead_trees}
\usage{
remove_dead_trees(tree.info, look.col, by.code = TRUE)
}
\arguments{
\item{tree.info}{dataframe}

\item{look.col}{}
}
\description{
Remove dead trees
need to handle at least two cases: i) remove by code ii) remove by mortality status
}
\author{
Istem Fer
}
