% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ca_compost.R
\name{sample_ca_compost_app_rate}
\alias{sample_ca_compost_app_rate}
\title{Sample California compost application rate by PFT family}
\usage{
sample_ca_compost_app_rate(pft_family, n = length(pft_family))
}
\arguments{
\item{pft_family}{character vector. PFT family for each draw. Must be
one of "annual", "perennial". Length 1 broadcasts to length n.}

\item{n}{integer. Number of draws. If pft_family has length > 1 then
n must equal length(pft_family).}
}
\value{
numeric vector of length n in tons per acre dry weight.
}
\description{
Draws n application rates (dry weight tons per acre) uniformly from
the per family envelope at ca_compost_app_rate_envelope.
}
\details{
Envelopes follow CDFA Healthy Soils Program white paper Table 2 plus
HSP 2026 Practice Guidelines plus USDA NRCS Conservation Practice
Standard 336 Soil Carbon Amendment (2022). Defaults: annual crops 3
to 8 t/ac dry, perennial crops 2 to 6 t/ac dry.
}
\examples{
sample_ca_compost_app_rate("annual", 5)
sample_ca_compost_app_rate(c("annual", "perennial", "annual"), 3)

}
