% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_utils.R
\name{sclass}
\alias{sclass}
\title{This function determines the soil class number based on the fraction of sand, clay, and silt}
\usage{
sclass(sandfrac, clayfrac)
}
\arguments{
\item{sandfrac, clayfrac}{numeric vectors with values in range 0 to 1.
Silt fraction is assumed to be the difference between (sand+clay) and 1}
}
\value{
vector of integers identifying textural class of each input layer.
 Possible values are 1 through 17; NB these are NOT the same class
 boundaries as the 12 USDA soil texture classes.
}
\description{
This function determines the soil class number based on the fraction of sand, clay, and silt
}
\examples{
sclass(0.3, 0.3)
}
