% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_utils.R
\name{soc2ocs}
\alias{soc2ocs}
\title{Convert soil organic carbon concentration to organic carbon stock}
\usage{
soc2ocs(soc_percent, bulk_density, thickness, coarse_fraction = 0)
}
\arguments{
\item{soc_percent}{soil organic carbon concentration (percent, 0-100)}

\item{bulk_density}{bulk density (g/cm3)}

\item{thickness}{layer thickness (cm)}

\item{coarse_fraction}{coarse fragment volume fraction (0-1, default = 0)}
}
\value{
organic carbon stock (kg/m2)
}
\description{
Convert soil organic carbon concentration to organic carbon stock
}
\examples{
soc2ocs(2.5, 1.3, 30, 0.15)

# Multiple soil layers
soc_pct <- c(3.2, 2.1, 1.8)
bd_g_cm3 <- c(1.2, 1.4, 1.5)
thickness_cm <- c(15, 15, 30)
coarse_fraction <- c(0.10, 0.20, 0.25)
soc2ocs(soc_pct, bd_g_cm3, thickness_cm, coarse_fraction)
}
\author{
Akash
}
