% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{soil.units}
\alias{soil.units}
\title{Get standard units for a soil variable}
\usage{
soil.units(varname = NA)
}
\arguments{
\item{varname}{}
}
\value{
character
}
\description{
Get standard units for a soil variable
}
\examples{
soil.units("soil_albedo")
}
