% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_utils.R
\name{soil_params}
\alias{soil_params}
\title{Estimate soil parameters from texture class or sand/silt/clay}
\usage{
soil_params(soil_type = NULL, sand = NULL, silt = NULL,
  clay = NULL, bulk = NULL)
}
\arguments{
\item{soil_type}{USDA Soil Class. See Details}

\item{sand}{percent sand}

\item{silt}{percent silt}

\item{clay}{percent clay}

\item{bulk}{soil bulk density (optional, kg m-3)}
}
\value{
list of soil hydraulic and thermal parameters
}
\description{
Estimate soil parameters from texture class or sand/silt/clay
}
\details{
* Specify _either_ soil_type or sand/silt/clay. soil_type will be ignored if sand/silt/clay is provided
* If only 2 out of sand/silt/clay are provided, it will be assumed they sum to 100%
* Valid soil class options: "Sand","Loamy sand","Sandy loam","Silt loam","Loam",
                            "Sandy clay loam","Silty clay loam","Clayey loam",
                            "Sandy clay","Silty clay","Clay","Peat","Bedrock",
                            "Silt","Heavy clay","Clayey sand","Clayey silt"
* Based on ED2/R-utils/soilutils.r
* Hydraulics based on Cosby et al 1984, using table 4 and equation 1 (which is incorrect it should be saturated moisture potential over moisture potential)
}
\examples{
sand <- c(0.3, 0.4, 0.5)
clay <- c(0.3, 0.3, 0.3)
soil_params(sand=sand,clay=clay)
}
