% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_params_ensemble.R
\name{soil_params_ensemble}
\alias{soil_params_ensemble}
\title{A function to estimate the soil parameters based on soil texture data and write the parameter paths into settings}
\usage{
soil_params_ensemble(
  settings,
  sand,
  clay,
  silt,
  outdir,
  write_into_settings = TRUE
)
}
\arguments{
\item{settings}{A multi-site settings}

\item{sand}{A data frame containing sand fraction in percentage from SoilGrids250m v2.0 with columns "Depth", "Quantile", "Siteid", and "Value"}

\item{clay}{A data frame containing clay fraction in percentage from SoilGrids250m v2.0 with columns "Depth", "Quantile", "Siteid", and "Value"}

\item{silt}{A data frame containing silt fraction in percentage from SoilGrids250m v2.0 with columns "Depth", "Quantile", "Siteid", and "Value"}

\item{outdir}{Provide the path to store the parameter files}

\item{write_into_settings}{Whether to write the path of parameter file into the setting. The default is TRUE}
}
\description{
A function to estimate the soil parameters based on soil texture data and write the parameter paths into settings
}
