% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soilgrids_SoilC_prep.R
\name{Soilgrids_SoilC_prep}
\alias{Soilgrids_SoilC_prep}
\title{Prepare Soilgrids SoilC data for the SDA workflow.}
\usage{
Soilgrids_SoilC_prep(
  site_info,
  start_date,
  end_date,
  time_points,
  outdir = NULL,
  export_csv = FALSE
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{start_date}{Start date of SDA workflow.}

\item{end_date}{End date of SDA workflow.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file will be stored.}

\item{export_csv}{Decide if we want to export the CSV file.}
}
\value{
A data frame containing AGB median and sd for each site and each time step.
}
\description{
Prepare Soilgrids SoilC data for the SDA workflow.
}
\author{
Dongchen Zhang
}
