% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids_texture_extract.R
\name{soilgrids_texture_extraction}
\alias{soilgrids_texture_extraction}
\alias{download_and_extraction}
\title{soilgrids_texture_extraction}
\usage{
download_and_extraction(base_data_url, site_info, outdir)
}
\arguments{
\item{site_info}{A data frame of site info containing the BETYdb site ID, 
site name, latitude, and longitude, e.g. 
(site_id, site_name, lat, lon)}

\item{outdir}{Provide the path to store the texture data file}

\item{verbose}{Provide progress feedback to the terminal? TRUE/FALSE}
}
\value{
a data frame containing the soil texture data with columns "Depth", "Quantile", "Siteid", and "Value"
}
\description{
soilgrids_texture_extraction function
A function to extract soil texture data for a single or group of 
lat/long locations based on user-defined site location from SoilGrids250m 
version 2.0 : https://soilgrids.org
}
\author{
Qianyu Li
}
