% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSSURGO_Query.R
\docType{data}
\name{SSURGO_API_MAX_AREA_M2}
\alias{SSURGO_API_MAX_AREA_M2}
\title{Maximum area for SSURGO API requests}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SSURGO_API_MAX_AREA_M2
}
\description{
Maximum area for SSURGO API requests
}
\keyword{datasets}
