% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSSURGO_Query.R
\name{ssurgo_mukeys_bbox}
\alias{ssurgo_mukeys_bbox}
\title{Get map unit keys (mukeys) from gSSURGO using a bounding box}
\usage{
ssurgo_mukeys_bbox(bbox)
}
\arguments{
\item{bbox}{Numeric vector of length 4: c(xmin, ymin, xmax, ymax) in WGS84 (EPSG:4326).
Features that intersect the bounding box are returned.}
}
\value{
Character vector of unique map unit keys (mukeys).
}
\description{
Queries the NRCS gSSURGO Web Feature Service to retrieve map unit keys
based on a bounding box spatial filter.
}
\details{
This function uses the NRCS SDM Data Access Web Feature Service:
\url{https://sdmdataaccess.nrcs.usda.gov/SpatialFilterHelp.htm}

The total extent of the bounding box cannot exceed 10,100,000,000 square
meters (~3,900 square miles). Use \code{ssurgo_mukeys_bigbbox()} for large
bounding boxes.
}
\examples{
\dontrun{
# Bounding box query
mukeys <- ssurgo_mukeys_bbox(bbox = c(-114.006, 32.1823, -113.806, 32.2823))
}
}
