% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.stringCode.R
\name{to.Tag}
\alias{to.Tag}
\title{to.Tag}
\usage{
to.Tag(SITE, PLOT, SUBPLOT, TAG = NULL)
}
\arguments{
\item{SITE, PLOT, SUBPLOT}{ignored}

\item{TAG}{string (or coercible to)}
}
\description{
to.Tag
}
