% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.stringCode.R
\name{to.TreeCode}
\alias{to.TreeCode}
\title{to.TreeCode}
\usage{
to.TreeCode(SITE, PLOT, SUBPLOT, TAG = NULL)
}
\arguments{
\item{SITE, PLOT, SUBPLOT, TAG}{strings (or coercible to)}
}
\description{
to.TreeCode
}
