% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPNET_flux_gapfill.R
\name{xgb_gapfill}
\alias{xgb_gapfill}
\title{xgb_gapfill}
\usage{
xgb_gapfill(
  site.dir,
  pred.var.dir,
  flux_var,
  outdir,
  vars_to_check = c("TA_F", "SW_IN_F", "SW_DIF", "LW_IN_JSB", "VPD_F", "PA_F", "P_F",
    "RH", "NETRAD", "PPFD_IN", "PPFD_DIF", "TS_F_MDS_1", "LW_IN_F", "TS_F_MDS_3",
    "TS_F_MDS_5", "SWC_F_MDS_1", "NDVI", "EVI", "sin_doy", "cos_doy", "SWC_F_MDS_3",
    "SWC_F_MDS_5"),
  n_sites = 40,
  missing_threshold = 40,
  nfolds = 10,
  nrounds = 200,
  nens = 25,
  nsamp = 0.95,
  params = list(objective = "reg:squarederror", eta = 0.13, max_depth = 13, subsample =
    0.9, colsample_bytree = 0.9),
  r2_threshold = 0.6,
  cores = max(1L, parallel::detectCores() - 1L),
  overwrite = TRUE
)
}
\arguments{
\item{flux_var}{character: the target flux type to be gap-filled. Currently
supported values are `"LE"` and `"NEE"`.}

\item{outdir}{character: the output directory where the exported CSV files
will be written.}

\item{nfolds}{numeric: number of folds used in CV, default is 10.}

\item{nrounds}{numeric: number of boosting rounds used in XGBoost training,
default is 200.}

\item{nens}{numeric: number of ensemble iterations, default is 25.}

\item{nsamp}{numeric: proportion of high-quality observations sampled within
each site for training in each ensemble, default is 0.95.}

\item{r2_threshold}{numeric: minimum average validation R2 required to
fit the final model and retain a site in the exported gap-filled output}

\item{cores}{numeric: how many CPUs to be used in the calculation}

\item{overwrite}{boolean: decide if existing output files should be overwritten,
the default is TRUE.}

\item{site_ids}{character: a character vector of site IDs to be modeled.}

\item{resimet.df}{data.frame: a data frame containing eddy-covariance
observations, QC flags, timestamps, and predictor variables.}

\item{sitecov.df}{data.frame: a data frame containing site-specific selected
predictor names, with columns including `Site_ID` and `filtered_variables`.}
}
\value{
list containing ensemble-level outputs. Each element corresponds to
one ensemble iteration and includes:
\describe{
  \item{cv_results}{a data frame of site-level average CV R2.}
  \item{filled_data}{a data frame of the full input dataset with merged gap-filled values.}
  \item{exported_data}{a data frame of filtered gap-filled results for well-performing sites only.}
  \item{folds_all}{a data frame of fold-level CV diagnostics across sites.}
}
}
\description{
This function performs site-level XGBoost gap-filling with CV
for eddy-covariance flux data. For each ensemble iteration, it samples
observed data(QC=0) within each site, fits site-specific XGBoost models
using k-fold CV, predicts missing or low-quality flux values,
and exports both gap-filled outputs and fold-level CV diagnostics.
}
\author{
Yang Gu
}
