% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{BADM_IC_process}
\alias{BADM_IC_process}
\title{BADM_IC_process}
\usage{
BADM_IC_process(settings, dir, overwrite = TRUE)
}
\arguments{
\item{settings}{Single PEcAn Settings object (list with 'run' element)}

\item{dir}{Output directory for storing IC netcdf file}

\item{overwrite}{Flag for overwriting the IC file.}
}
\value{
a list of paths to generated and stored IC files.
}
\description{
Process initial conditions for a single site using BADM data.
This function accepts only single Settings objects.
For processing multiple sites, use a higher-level function.
}
