% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{BADM_IC_process}
\alias{BADM_IC_process}
\title{BADM_IC_process}
\usage{
BADM_IC_process(settings, dir, overwrite = TRUE)
}
\arguments{
\item{settings}{pecan xml settings}

\item{dir}{output dir which you want to store the IC netcdf file}

\item{overwrite}{Flag for overwriting the IC file.}
}
\value{
a list of paths to generated and stored IC files.
}
\description{
BADM_IC_process
}
