% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort2pool.R
\name{cohort2pool}
\alias{cohort2pool}
\title{cohort2pool function
Calculates total biomass using veg cohort file.}
\usage{
cohort2pool(veg_file, allom_param = NULL)
}
\arguments{
\item{veg_file}{path to standard cohort veg_file}

\item{allom_param}{parameters for allometric equation, a and b. Based on base-10 log-log linear model (power law)}
}
\description{
cohort2pool function
Calculates total biomass using veg cohort file.
}
\examples{
\dontrun{
veg_file <- "~/downloads/FFT_site_1-25665/FFT.2008.veg.rds"
cohort2pool(veg_File = veg_file, allom_param = NULL)
}
}
\author{
Saloni Shah
}
