% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort2poll.R
\name{cohort2pool}
\alias{cohort2pool}
\title{cohort2pool}
\usage{
cohort2pool(veg_file, allom_param = NULL)
}
\arguments{
\item{veg_file}{standard cohort veg_file}

\item{allom_param}{parameters for allometric equation, a and b}
}
\author{
Saloni Shah
}
