% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{EPA_ecoregion_finder}
\alias{EPA_ecoregion_finder}
\title{EPA_ecoregion_finder}
\usage{
EPA_ecoregion_finder(Lat, Lon, folder.path = NULL)
}
\arguments{
\item{Lat}{numeric latitude}

\item{Lon}{numeric longitude}

\item{folder.path}{path to the directory where you store the shape files of L1 and L2 ecoregion maps.}
}
\value{
a dataframe with codes corresponding to level1 and level2 codes as two columns
}
\description{
This function is designed to find the level1 and level2 code ecoregions for a given lat and long. 
You can learn more about ecoregions here: \url{https://www.epa.gov/eco-research/ecoregions}.
}
