% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMIS-ET.R
\name{extract_cimis_dates}
\alias{extract_cimis_dates}
\title{Extract CIMIS reference ET for multiple dates}
\usage{
extract_cimis_dates(design_points, dates, .progress = FALSE, ...)
}
\arguments{
\item{design_points}{`data.frame` of design points with columns
`location_id`, `lat`, and `lon`}

\item{dates}{Sequence of dates for which to extract data}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on
a basic progress bar, use a string to give it a name, or see
\link[purrr]{progress_bars} for more details.}

\item{...}{Additional arguments to [extract_cimis_date()]}
}
\value{
`design_points` `data.frame` extended with ETref data for all dates.
}
\description{
Extract CIMIS reference ET for multiple dates
}
