% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_FIA.R, R/write_veg.R
\name{extract_FIA}
\alias{extract_FIA}
\alias{extract_FIA}
\alias{write_veg}
\title{extract_FIA}
\usage{
extract_FIA(lon, lat, start_date, gridres = 0.075, dbparms)

write_veg(outfolder, start_date, end_date, veg_info, overwrite = FALSE, ...)
}
