% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_FIA.R
\name{extract_FIA}
\alias{extract_FIA}
\title{extract_FIA}
\usage{
extract_FIA(lon, lat, start_date, end_date, gridres = 0.075, dbparms, ...)
}
\arguments{
\item{lon}{site longitude}

\item{lat}{site latitude}

\item{start_date}{"YYYY-MM-DD"}

\item{end_date}{"YYYY-MM-DD"}

\item{gridres}{taken from input_veg, DEFAULT = 0.075}

\item{dbparms}{taken from settings object}

\item{...}{Additional parameters}
}
\description{
extract_FIA
}
\author{
Istem Fer
}
