% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NEON_veg.R
\name{extract_NEON_veg}
\alias{extract_NEON_veg}
\title{extract_NEON_veg}
\usage{
extract_NEON_veg(new_site, start_date, end_date, ...)
}
\arguments{
\item{new_site}{object passed from get_veg_module that includes site lat, lon, id, and name}

\item{start_date}{"YYYY-MM-DD", used to download NEON datasets for desired time period}

\item{end_date}{"YYYY_MM_DD", used to download NEON datasets for desired time period}

\item{...}{Additional parameters}
}
\value{
veg_info object to be passed to extract_veg within ic_process
}
\description{
extract_NEON_veg
}
\examples{
start_date = as.Date("2020-01-01") end_date = as.Date("2021-09-01")
}
\author{
Alexis Helgeson and Michael Dietze
}
