% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NEON_veg.R
\name{extract_NEON_veg}
\alias{extract_NEON_veg}
\title{extract_NEON_veg}
\usage{
extract_NEON_veg(lon, lat, start_date, end_date, store_dir, ...)
}
\arguments{
\item{lon}{site longitude coordinate passed from get_veg_module}

\item{lat}{site latitude coordinate passed from get_veg_module}

\item{startdate}{"YYYY-MM-DD", used to download NEON datasets for desired time period}

\item{enddate}{"YYYY_MM_DD", used to download NEON datasets for desired time period}

\item{store_dir}{location to store downloaded NEON files}

\item{...}{Additional parameters}
}
\value{
veg_info object to be passed to extract_veg within ic_process
}
\description{
extract_NEON_veg
}
\examples{
start_date = as.Date("2020-01-01") 
end_date = as.Date("2021-09-01")
}
\author{
Alexis Helgeson and Michael Dietze
}
