% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NEON_veg.R
\name{extract_NEON_veg}
\alias{extract_NEON_veg}
\title{extract_NEON_veg}
\usage{
extract_NEON_veg(
  lon,
  lat,
  start_date,
  end_date,
  store_dir,
  neonsites = NULL,
  ...
)
}
\arguments{
\item{lon}{site longitude, passed from ic_process}

\item{lat}{site latitude, passed from ic_process}

\item{start_date}{"YYYY-MM-DD", used to download NEON datasets for desired time period}

\item{end_date}{"YYYY_MM_DD", used to download NEON datasets for desired time period}

\item{store_dir}{location where you want to store downloaded NEON files}

\item{neonsites}{prepared datasets table from NEON using neonstore::neon_sites(api = "https://data.neonscience.org/api/v0", .token = Sys.getenv("NEON_TOKEN"))}

\item{...}{Additional parameters}
}
\value{
veg_info object to be passed to extract_veg within ic_process
}
\description{
extract_NEON_veg
}
\examples{
start_date = as.Date("2020-01-01") 
end_date = as.Date("2021-09-01")
}
\author{
Alexis Helgeson and Michael Dietze
}
