% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{extract_soil_gssurgo}
\alias{extract_soil_gssurgo}
\title{Extract soil data from gssurgo}
\usage{
extract_soil_gssurgo(
  outdir,
  lat,
  lon,
  size = 1,
  grid_size = 3,
  grid_spacing = 10,
  depths = c(0.15, 0.3, 0.6)
)
}
\arguments{
\item{outdir}{Output directory for writing down the netcdf file}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{size}{Ensemble size}

\item{grid_size}{Size of the spatial sampling grid (default: 3)}

\item{grid_spacing}{Spacing between grid cells in meters (default: 10)}

\item{depths}{Standard set of soil depths in m to create the ensemble of soil profiles with.}
}
\value{
It returns the address for the generated soil netcdf file
}
\description{
Extract soil data from gssurgo
}
\examples{
\dontrun{
   outdir  <- "~/paleon/envTest"
   lat     <- 40
   lon     <- -80
   PEcAn.data.land::extract_soil_gssurgo(outdir, lat, lon)
}
}
\author{
Hamze Dokoohaki and Akash
}
