% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{extract_soil_gssurgo}
\alias{extract_soil_gssurgo}
\title{Extract soil data}
\usage{
extract_soil_gssurgo(outdir, lat, lon)
}
\arguments{
\item{outdir}{}

\item{lat}{}

\item{lon}{}
}
\description{
Extract soil data
}
\examples{
outdir  <- "~/paleon/envTest"
lat     <- 40
lon     <- -80
\dontrun{
   PEcAn.data.land::extract_soil_gssurgo(outdir,lat,lon)
}
}
\author{
Hamze Dokoohaki
}
