% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_nc.R
\name{extract_soil_nc}
\alias{extract_soil_nc}
\title{Extract soil data}
\usage{
extract_soil_nc(in.file, outdir, lat, lon)
}
\arguments{
\item{in.file}{}

\item{outdir}{}

\item{lat}{}

\item{lon}{}
}
\value{
path to netCDF file containing extracted data
}
\description{
Extract soil data
}
\examples{
\dontrun{
in.file <- "~/paleon/env_paleon/soil/paleon_soil.nc"
outdir  <- "~/paleon/envTest"
lat     <- 40
lon     <- -80
PEcAn.data.land::extract_soil_nc(in.file,outdir,lat,lon)
}
}
