% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_veg.R
\name{extract_veg}
\alias{extract_veg}
\title{load_veg}
\usage{
extract_veg(new_site, start_date, end_date, source, gridres,
  format_name = NULL, machine_host, dbparms, outfolder,
  overwrite = FALSE, ...)
}
\description{
Function queries a DB to extract veg info downstream
}
\author{
Istem Fer
}
