% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizer_composition_data}
\alias{fertilizer_composition_data}
\title{Fertilizer Nutrient Composition Table}
\format{

}
\source{
https://github.com/swat-model/swatplus

DayCent model default parameter file: `omad.100` obtained from the Soil Carbon Solutions Center, https://www.soilcarbonsolutionscenter.com
}
\usage{
fertilizer_composition_data
}
\description{
A dataset of fertilizer and organic matter addition types 
and their nitrogen and carbon composition, based on the SWAT model's
`fertilizer.frt` table and DayCent model defaults for organic matter
C:N ratio parameters.
}
\details{

}
\keyword{datasets}
