% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia2ED.R
\name{fia.to.psscss}
\alias{fia.to.psscss}
\title{Create pss/css files based on data in the fia database}
\usage{
fia.to.psscss(
  settings,
  lat = as.numeric(settings$run$site$lat),
  lon = as.numeric(settings$run$site$lon),
  year = lubridate::year(settings$run$start.date),
  gridres = 0.075,
  min.year = year - 5,
  max.year = year + 5,
  overwrite = FALSE
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{lat, lon}{site location in decimal degrees.
Defults to values passed in `settings`.}

\item{year}{defaults to year of start date passed in settings}

\item{gridres}{grid resolution in degrees}

\item{min.year, max.year}{limits on years of FIA data to look for}

\item{overwrite}{logical: regenerate files already in the database?}
}
\value{
modified settings, invisibly
}
\description{
Create pss/css files based on data in the fia database
}
\author{
Mike Dietze, Rob Kooper, Ryan Kelly
}
