% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fia2ED.R
\name{fia.to.psscss}
\alias{fia.to.psscss}
\title{fia.to.psscss}
\usage{
fia.to.psscss(settings, gridres = 0.075)
}
\arguments{
\item{create}{pss/css files based on data in the fia database}
}
\value{
nothing
}
\description{
convert x into a table
}
\author{
Mike Dietze, Rob Kooper
}

