% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataONE_doi_download.R
\name{format_identifier}
\alias{format_identifier}
\title{format_identifier}
\usage{
format_identifier(id)
}
\arguments{
\item{id}{the doi or other identifier linked to the package in DataONE}
}
\value{
returns the id in the proper format for querying the DataONE Federation (using solrQuery syntax)
}
\description{
This function is for formatting purposes. It simply inserts the doi or id that the user wishes to query into Solr format so that it is compatible with the dataoneR query functionality in the PEcAn function
}
\author{
Liam P Burke, \email{lpburke@bu.edu}
}
