% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.soil}
\alias{get.soil}
\title{get.soil}
\usage{
get.soil(lat, lon, soil.nc = soil.nc)
}
\arguments{
\item{lat}{}

\item{lon}{}

\item{soil.nc}{netCDF file with soil data}
}
\value{
usda soil class
}
\description{
Get Soil
}
\author{
David LeBauer
}

