% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.soil.R
\name{get.soil}
\alias{get.soil}
\title{get.soil}
\usage{
get.soil(lat, lon, soil.nc = soil.nc)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{soil.nc}{netCDFe file with soil data}
}
\value{
usda soil class
}
\description{
Get Soil
}
\author{
David LeBauer
}

