% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSSURGO_Query.R
\name{gSSURGO.Query}
\alias{gSSURGO.Query}
\title{This function queries the gSSURGO database for a series of map unit keys}
\usage{
gSSURGO.Query(mukeys = 2747727)
}
\arguments{
\item{mukeys}{map unit key from gssurgo}
}
\value{
a dataframe with soil properties. units can be looked up from database documentation
}
\description{
This function queries the gSSURGO database for a series of map unit keys
}
\details{
Full documention of available tables and their relationships can be found here \url{www.sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx}
There have been occasions where NRCS made some minor changes to the structure of the API which this code is where those changes need
to be implemneted here.
}
