% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_process.R
\name{ic_process}
\alias{ic_process}
\title{ic_process}
\usage{
ic_process(settings, input, dir, overwrite = FALSE)
}
\arguments{
\item{settings}{pecan settings list}

\item{input}{Taken from settings$run$inputs. This should include id, path, and source}

\item{dir}{settings$database$dbfiles}

\item{overwrite}{Default = FALSE. whether to force ic_process to proceed}
}
\description{
ic_process
}
\author{
Istem Fer, Hamze Dokoohaki
}
