% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_process.R
\name{ic_process}
\alias{ic_process}
\title{ic_process}
\usage{
ic_process(settings, input, dir, overwrite = FALSE)
}
\arguments{
\item{settings}{pecan settings list}

\item{overwrite}{whether to force ic_process to proceed}

\item{dbfiles}{where to write files}
}
\description{
ic_process
}
\author{
Istem Fer
}
