% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InventoryGrowthFusion.R
\name{InventoryGrowthFusion}
\alias{InventoryGrowthFusion}
\title{InventoryGrowthFusion}
\usage{
InventoryGrowthFusion(data, n.iter, random = TRUE)
}
\arguments{
\item{data}{list of data inputs}

\item{random}{= whether or not to include random effects}
}
\value{
an mcmc.list object
}
\description{
this code fuses forest inventory data with tree growth data (tree ring or dendrometer band)
for the same plots. Code is a rewrite of Clark et al 2007 Ecol Appl into JAGS
}
\note{
Requires JAGS
}

