% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iscn_soc}
\alias{iscn_soc}
\title{Soil organic carbon (SOC) density based on eco-region level 2 code from the ISCN database.}
\format{
## `iscn_soc`
A data frame with 200 rows and 43 columns:
\describe{
  \item{rows}{1 to 200 ensemble members}
  \item{columns}{43 level 2 ecoregion codes across North America}
}
}
\source{
https://iscn.fluxdata.org/wp-content/uploads/sites/23/2019/05/ISCN_ALL_DATA_DATASET_1-1.xlsx
}
\usage{
iscn_soc
}
\description{
Contains 200 ensemble SOC data from 43 level 2 eco-regions across North America.
Variable include SOC densities in g/cm2.
}
\keyword{datasets}
