% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_veg.R
\name{load_veg}
\alias{load_veg}
\title{load_veg}
\usage{
load_veg(
  new_site,
  start_date,
  end_date,
  source_id,
  source,
  icmeta = NULL,
  format_name = NULL,
  machine_host,
  dbparms,
  outfolder,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{new_site}{list passed to `load_data`}

\item{start_date, end_date}{date range to look up}

\item{source_id}{input id to look up in DB}

\item{source}{name of data source (used in file naming)}

\item{icmeta}{metadata for initial conditions}

\item{format_name}{file format to look for}

\item{machine_host}{hostname of machine where the data lives}

\item{dbparms}{parameters to use when opening connection to database}

\item{outfolder}{path to write results}

\item{overwrite}{Logical: replace existing files? NOTE: Currently ignored!}

\item{...}{Additional arguments, currently ignored}
}
\description{
uses `PEcAn.benchmark::load_data()` to get veg data
}
\author{
Istem Fer
}
