% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pft.R
\name{match_pft}
\alias{match_pft}
\title{Match model PFTs}
\usage{
match_pft(
  bety_species_id,
  pfts,
  query = NULL,
  con = NULL,
  allow_missing = FALSE
)
}
\arguments{
\item{bety_species_id}{vector of BETYdb species IDs}

\item{pfts}{settings$pfts.  List of pfts with database matching based on name}

\item{con}{database connection, if NULL use traits package}

\item{allow_missing}{flag to indicate that settings file does not need to match exactly}
}
\value{
table of BETYdb PFT IDs matched to species IDs
}
\description{
Matches BETYdb species IDs to model-specific PFTs
}
\author{
Mike Dietze, Istem Fer
}
