% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_species_id.R
\name{match_species_id}
\alias{match_species_id}
\title{Match BETY species ID.}
\usage{
match_species_id(input_codes, format_name, betydb = NULL,
  custom_translation_path = NULL, ...)
}
\arguments{
\item{input_codes}{Character vector of species codes}

\item{format_name}{Species code format name (character, length 1)}

\item{betydb}{\code{dplyr} \code{src} object containing BETY connection}

\item{custom_translation_path}{File path to translation}
}
\value{
\code{data.frame} containing the following columns:
\describe{
 \item{\code{input_code}}{Character provided as input}
 \item{\code{bety_species_id}}{Big integer species ID, unique and specific to BETY}
 \item{\code{genus}}{Genus part of Latin name, from BETY}
 \item{\code{species}}{Species part of Latin name, from BETY}
}
}
\description{
Parses species codes in input data and matches them with the BETY species ID.
}
\examples{
betydb <- dplyr::src_postgres(dbname = 'bety',
                       user = 'bety',
                       password = 'bety',
                       host = 'localhost')
input_codes <- c('ACRU', 'PIMA', 'TSCA')
format_name <- 'usda'
match_species_id(input_codes = input_codes,
                 format_name = format_name,
                 betydb = betydb)
                 
}
\author{
Alexey Shiklomanov <ashiklom@bu.edu>
}

