% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_species_id.R, R/match_species_id_OLD.R
\name{match_species_id}
\alias{match_species_id}
\title{Match BETY species ID.}
\usage{
match_species_id(input_codes, format_name = "custom", bety = NULL,
  translation_table = NULL, ...)

match_species_id(input_codes, format_name = "custom", bety = NULL,
  translation_table = NULL, ...)
}
\arguments{
\item{input_codes}{Character vector of species codes}

\item{format_name}{Species code format name (see details)}

\item{bety}{\code{dplyr} \code{src} object containing BETY connection}

\item{translation_table}{Data frame with custom translation table (see details).}

\item{input_codes}{Character vector of species codes}

\item{format_name}{Species code format name (see details)}

\item{bety}{\code{dplyr} \code{src} object containing BETY connection}

\item{translation_table}{Data frame with custom translation table (see details).}
}
\value{
\code{data.frame} containing the following columns:
\describe{
 \item{\code{input_code}}{Character provided as input}
 \item{\code{bety_species_id}}{Big integer species ID, unique and specific to BETY}
 \item{\code{genus}}{Genus part of Latin name, from BETY}
 \item{\code{species}}{Species part of Latin name, from BETY}
}

\code{data.frame} containing the following columns:
\describe{
 \item{\code{input_code}}{Character provided as input}
 \item{\code{bety_species_id}}{Big integer species ID, unique and specific to BETY}
 \item{\code{genus}}{Genus part of Latin name, from BETY}
 \item{\code{species}}{Species part of Latin name, from BETY}
}
}
\description{
Parses species codes in input data and matches them with the BETY species ID.

Parses species codes in input data and matches them with the BETY species ID.
}
\details{
\code{format_name} can be one of the following:
\describe{
 \item{\code{usda}}{USDA Plants database symbol (e.g. QURU, TSCA)}
 \item{\code{fia}}{FIA species code}
 \item{\code{latin_name}}{Scientific name, as "Genus species"; must match exactly and unambiguously to \code{scientificname} field in BETY}
 \item{\code{custom}}{A data frame matching BETY IDs (column name \code{bety_species_id}) to input codes (column name \code{input_code}). This data frame must be passed via the \code{translation_table} argument.}
}

\code{format_name} can be one of the following:
\describe{
 \item{\code{usda}}{USDA Plants database symbol (e.g. QURU, TSCA)}
 \item{\code{fia}}{FIA species code}
 \item{\code{latin_name}}{Scientific name, as "Genus species"; must match exactly and unambiguously to \code{scientificname} field in BETY}
 \item{\code{custom}}{A data frame matching BETY IDs (column name \code{bety_species_id}) to input codes (column name \code{input_code}). This data frame must be passed via the \code{translation_table} argument.}
}
}
\examples{
bety <- dplyr::src_postgres(dbname = 'bety',
                       user = 'bety',
                       password = 'bety',
                       host = 'localhost')
input_codes <- c('ACRU', 'PIMA', 'TSCA')
format_name <- 'usda'
match_species_id(input_codes = input_codes,
                 format_name = format_name,
                 bety = bety)

}
\author{
Alexey Shiklomanov <ashiklom@bu.edu>, Istem Fer

Alexey Shiklomanov <ashiklom@bu.edu>, Istem Fer
}
