% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_utils.R
\name{mpot2smoist}
\alias{mpot2smoist}
\title{Convert a matric potential to a soil moisture}
\usage{
mpot2smoist(
  mpot,
  soil_water_potential_at_saturation,
  soil_hydraulic_b,
  volume_fraction_of_water_in_soil_at_saturation
)
}
\arguments{
\item{mpot}{water potential (cm H2O)}

\item{soil_water_potential_at_saturation}{water potential when soil is saturated (cm H2O)}

\item{soil_hydraulic_b}{pore-size distribution parameter for Campbell (1974) water content model}

\item{volume_fraction_of_water_in_soil_at_saturation}{VSWC when soil is saturated (numeric in range 0-1)}
}
\value{
volumetric soil water content
}
\description{
Convert a matric potential to a soil moisture
}
