% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{netcdf.writer.BADM}
\alias{netcdf.writer.BADM}
\title{netcdf.writer.BADAM}
\usage{
netcdf.writer.BADM(lat, long, siteid, outdir, ens)
}
\arguments{
\item{lat}{numeric latitude}

\item{long}{numeric longitude}

\item{siteid}{site id as a string}

\item{outdir}{output dir which you want to store the IC netcdf file}

\item{ens}{ensemble members, passed on to `pool_ic_list2netcdf`}
}
\value{
a dataframe with file, host, mimetype, formatname, startdate, enddate and dbfile.name columns
}
\description{
netcdf.writer.BADAM
}
