% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSSURGO_Query.R
\name{parse_mukey_response}
\alias{parse_mukey_response}
\title{Parse responses from the mukey WFS service}
\usage{
parse_mukey_response(resp)
}
\arguments{
\item{resp}{`httr2` response object from SSURGO mukey WFS API}
}
\value{
character vector of mukeys
}
\description{
Parse responses from the mukey WFS service
}
