% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot2AGB.R
\name{plot2AGB}
\alias{plot2AGB}
\title{plot2AGB}
\usage{
plot2AGB(combined, out, outfolder, allom.stats, unit.conv = 0.02)
}
\arguments{
\item{combined}{data frame merging plot inventory and tree ring data}

\item{out}{MCMC samples for diameter (sample x tree)}

\item{outfolder}{output folder for graphs & data}

\item{unit.conv}{area conversion from sum(kg/tree) to kg/area}
}
\description{
convert composite ring & census data into AGB
}
\author{
Mike Dietze \email{dietze@bu.edu}
}

