% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_SOILGRID_Utilities.R
\name{preprocess_soilgrids_data}
\alias{preprocess_soilgrids_data}
\title{Preprocess SoilGrids data}
\usage{
preprocess_soilgrids_data(
  soil_data,
  default_soilC = 5,
  default_uncertainty = 0.2,
  verbose = FALSE
)
}
\arguments{
\item{soil_data}{Raw soil carbon data from soilgrids_soilC_extract}

\item{default_soilC}{Default soil carbon value in kg/m² to use when data is missing}

\item{default_uncertainty}{Default uncertainty as fraction to use when data is missing}

\item{verbose}{Print detailed progress information to the terminal? TRUE/FALSE}
}
\value{
Processed soil carbon data
}
\description{
Preprocess SoilGrids data
}
