% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BADM_Utilities.R
\name{Read.IC.info.BADM}
\alias{Read.IC.info.BADM}
\title{Read.IC.info.BADM}
\usage{
Read.IC.info.BADM(lat, long)
}
\arguments{
\item{lat}{numeric latitude}

\item{long}{numeric longitude}
}
\value{
a dataframe with 8 columns of Site, Variable, Date, Organ, PlantWIni (Initial plant biomass, type of biomass can be found in the Variable and Organ), SoilIni (which shows the initial soil C), LitterIni, RootIni.
}
\description{
This function returns a dataframe of plant biomass, root and soil carbon for a set of lat and long coordinates.
This function first finds the level1 and level2 ecoregions  for the given coordinates, and then tries to filter BADM database for those eco-regions. 
If no data found in the BADM database for the given lat/longs eco-regions, then all the data in the database will be used to return the initial condition.
All the variables are also converted to kg/m^2.
}
\examples{
\dontrun{
  badm_test <- Read.IC.info.BADM(45.805925,-90.07961)
}
}
