% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Tuscon.R
\name{Read_Tucson}
\alias{Read_Tucson}
\title{Read_Tucson}
\usage{
Read_Tucson(folder)
}
\description{
wrapper around read.tucson that loads a whole directory of tree ring files 
and calls a 'clean' function that removes redundant records 
(WinDendro can sometimes create duplicate records when editing)
}

