% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_params_ensemble.R
\name{soil_params_ensemble_soilgrids}
\alias{soil_params_ensemble_soilgrids}
\title{A function to estimate the soil parameters based on SoilGrids soil texture
 data and write the parameter paths into settings}
\usage{
soil_params_ensemble_soilgrids(
  settings,
  sand,
  clay,
  silt,
  outdir,
  write_into_settings = TRUE
)
}
\arguments{
\item{settings}{A multi-site settings}

\item{sand, clay, silt}{Data frames containing fraction in percentage from SoilGrids250m
v2.0, each with columns "Depth", "Quantile", "Siteid", and "Value"}

\item{outdir}{Provide the path to store the parameter files}

\item{write_into_settings}{Whether to write the path of parameter file into
the setting. The default is TRUE}
}
\value{
Ensemble soil parameter files defined in outdir and file paths in xml file
}
\description{
A function to estimate the soil parameters based on SoilGrids soil texture
 data and write the parameter paths into settings
}
\examples{
\dontrun{

outdir <- "/projectnb/dietzelab/Cherry/SoilGrids_texture/39NEON"
# each file contains percent salt, silt, or clay
sand <- readRDS("/path/to/SoilGrids_texture/sand_percent.rds")
clay <- readRDS("/path/to/SoilGrids_texture/clay_percent.rds")
silt <- readRDS("/path/to/SoilGrids_texture/silt_percent.rds")
settings <-read.settings("/path/to/pecan_monthly_SDA_soilwater.xml")
soil_params_ensemble_soilgrids(settings,sand,clay,silt,outdir)
}

}
\author{
Qianyu Li
}
