% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_process.R
\name{soil_process}
\alias{soil_process}
\title{Module for managing soil texture extraction}
\usage{
soil_process(settings, input, dbfiles, overwrite = FALSE, run.local = TRUE)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{input}{PEcAn input list}

\item{dbfiles}{directory to write database files}

\item{overwrite}{overwrite previous results (boolean)}

\item{run.local}{logical: Run only on the current machine?
If FALSE, runs on `settings$host` (which might turn out to be the current machine)}
}
\value{
path to soil file
}
\description{
Module for managing soil texture extraction
}
