% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids_texture_extract.R
\name{soilgrids_texture_extraction}
\alias{soilgrids_texture_extraction}
\title{soilgrids_texture_extraction}
\usage{
soilgrids_texture_extraction(
  data_paths,
  site_info,
  outdir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_paths}{A list containing the data (either virtual raster files or local file folders) and output path for all types of SoilGrids texture data
e.g. data_paths <- list(
 list(
    url = "/vsicurl?max_retry=30&retry_delay=60&list_dir=no&url=https://files.isric.org/soilgrids/latest/data/sand/sand_",
    local = NULL,
    save_path = paste0(outdir, "sand_percent.rds")),
 list(
    url = "/vsicurl?max_retry=30&retry_delay=60&list_dir=no&url=https://files.isric.org/soilgrids/latest/data/clay/clay_",
    local = NULL,
    save_path = paste0(outdir, "clay_percent.rds")),
 list(
    url = "/vsicurl?max_retry=30&retry_delay=60&list_dir=no&url=https://files.isric.org/soilgrids/latest/data/silt/silt_",
    local = NULL,
    save_path = paste0(outdir, "silt_percent.rds")))}

\item{site_info}{A data frame of site info containing the BETYdb site ID, 
site name, latitude, and longitude, e.g. 
(site_id, lat, lon)}

\item{outdir}{Provide the path to store the texture data file}

\item{verbose}{Provide progress feedback to the terminal? TRUE/FALSE}
}
\value{
a data frame containing the soil texture data with columns "Depth", "Quantile", "Siteid", and "Value"
}
\description{
soilgrids_texture_extraction function
A function to extract and save three types of soil texture data in parallel for a single or group of 
lat/long locations based on user-defined site location from SoilGrids250m 
version 2.0 : https://soilgrids.org
}
\author{
Qianyu Li
}
